requirements.txt
geopandas==0.5.1
numpy==1.16.4
Flask==1.1.1
pandas==0.24.2
feather_format==0.4.0
matplotlib==3.1.1
geopy==1.20.0
WTForms==2.2.1
seaborn==0.9.0
Cython==0.29.13
pyOpenSSL==19.0.0
cryptography==2.7
protobuf==3.9.1
ipaddr==2.2.0
lxml==4.4.1
ordereddict==1.1
wincertstore==0.2
gunicorn==19.9.0
numpy==1.16.4
scikit-learn==0.21.3
scipy==1.3.1
